+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Dialog的OK/Cancel按钮会根据操作系统选择不同的排列顺序
 
 ## [beta-0.6.0] - 2021-11-28
 ### Added
@@ -31,7 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - query the number of DBs without using the 'CONFIG GET' command
 - replace Label with Jlabel
 
-### Removed
 
 [Unreleased]: https://github.com/MattMin/a-redis/compare/beta-0.6.0...HEAD
 [beta-0.6.0]: https://github.com/MattMin/a-redis/releases/tag/beta-0.6.0